refactor(170): consolidate routes and ensure immutable config snapshots
Migrate four route groups from main.py to domain routers for better separation of concerns:
- Retention config (GET/PATCH /api/admin/retention) → admin_router.py
- Job definitions alias (GET /job-definitions) → jobs_router.py
- Verification key and docs routes → system_router.py

Ensure VaultConfigSnapshot is used everywhere:
- vault_router.py line 87: convert config to frozen snapshot on update
- vault_router.py line 120-128: use VaultConfigSnapshot for test connection

Benefits:
- Reduced main.py sprawl (now 1006 lines vs 1049)
- Immutable config pattern prevents DetachedInstanceError (D-05)
- Better code organization with domain-specific routers
- Consistent use of frozen snapshots across EE services
